make mauncher way faster to launch
So, this commit is kind of stupid. It increases complexity a ton.
Howver, on my laptop, with TLP running and clockig my CPU down to 900MHz
to save power, just the time it takes between g_application_run and the
"activate" callback was unacceptably long.

This commit adds a mauncher daemon, which always runs in the background
and is ready to open a window in an instant whenever the mauncher client
connects. In order to keep the same interface, the daemon and client are
the same binary, and the daemon is automatically spawned if the client
detects that no daemon is running.